services/flight-ingestion/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
confluent-kafka==2.5.3
python-jose==3.3.0
strawberry-graphql==0.243.0
prometheus-fastapi-instrumentator==7.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
